Avoid implementations with x86-specific assembly code
For Wasm builds, we should avoid x86-specific code. The normal Wasm build has a `generic` target but to use the x86 SIMD code we must use the `x86_64` target that enables all of this code. We hide it behind `CONFIG_RUNTIME_CPU_DETECT` since this is unnecessary in Wasm. Note that not all of the alternate implementations are complete and/or correct (e.g. some just return 0).